move larger functions to cc file
